fix scripts
